requirements.txt
# Image Triangulation Python Library Requirements

numpy>=1.21.0,<2.0.0
scipy>=1.7.0
Pillow>=8.3.0
matplotlib>=3.4.0,<4.0.0




